Large change set
Forgot to check these in in small increments while coding. One of those
I did it all at the same time. Changes the buttons and general look of
the menu. Included a header and a new back button. Properly fixed the
jumping glitch. Made the top button only appear when there are
scrollbars with a new animation, also added in an option to disable it
completely. There is also a new property called home, if you give it a
url it places a home button on the menu, if the current page and home
url are the same the button is hidden.