Update to better handle small screens and resizing
This update allows FooNav to work a lot better on small screens and
during resize operations. On small screens the size of the menu is set
and the items become scrollable, also any long text items become
wrapped. FooNav will now check it's size on window resize and act
according.